+++ src/components/header.js
@@ -1,12 +1,6 @@
-import { Link } from 'gatsby';
 import React from 'react';
-import { Navbar, Nav } from 'react-bootstrap';
 import '../styles/header.scss';
-
-const headerActiveStyle = {
-  textDecoration: 'underline',
-  textUnderlinePosition: 'under',
-};
+import NavBar from './navBar';
 
 const Header = () => (
   <header
@@ -14,46 +8,7 @@ const Header = () => (
       background: `white`,
     }}
   >
-    <Navbar bg='transparent'>
-      <Navbar.Brand>
-        <Link
-          to='/'
-          style={{
-            color: 'black',
-            textDecoration: 'none',
-          }}
-        >
-          AS
-        </Link>
-      </Navbar.Brand>
-
-      <Nav className='ml-auto'>
-        <Link to='/' className='header-link' activeStyle={headerActiveStyle}>
-          Home
-        </Link>
-        <Link
-          to='/about'
-          className='header-link'
-          activeStyle={headerActiveStyle}
-        >
-          About
-        </Link>
-        <Link
-          to='/skills'
-          activeStyle={headerActiveStyle}
-          className='header-link'
-        >
-          Skills
-        </Link>
-        <Link
-          to='/contact'
-          activeStyle={headerActiveStyle}
-          className='header-link'
-        >
-          Contact
-        </Link>
-      </Nav>
-    </Navbar>
+    <NavBar></NavBar>
   </header>
 );
