+++ src/styles/layout.scss
@@ -55,3 +55,7 @@ $bp-medium: 768px;
   justify-content: center;
   align-items: center;
 }
+
+.justify-end{
+    justify-content: flex-end;
+}
